non-functional: Simply parsing using macros

+++ src/model/control_sequence/action.rs
@@ -48,11 +48,11 @@ pub enum Action {
     SaveCursor,
     RestoreCursor,
 
-    HorizontalMove(isize),
+    HorizontalMove(u32),
 
     VerticalPos(isize),
 
-    DA1(usize),
+    DA1(u32),
 
     WindowOps(u8, usize, usize),
